fix_session: fix_session_deactivate: fix race condition
A race condition between the main thread and any FixSession
object's socket thread could be encountered when the server
is forcibly exited with CTRL-C and cause a segmentation fault.

The root cause of the race is an incorrect use of the return
value of pthread_equal in fix_session_deactivate which meant
that the caller of fix_session_deactivate wasn't reaching any
of the pthread_join calls to waiting for the Tx, Rx, and socket
threads to end.

Here is a description of the race:

1. SIGINT sent to server
2. Server calls fix_session_free() from the main thread,
   which subsequently calls fix_session_deactivate()
3. Session y deactivated from main thread. Incorrect use
   of pthread_equal means that the deactivation doesn't
   pthread_join on the session Tx, Rx, and socket threads
4. Session deactivation from the main thread completes and
   closes the session's socket
5. At about the same time, session y's socket thread,
   which has been blocked on recv, detects that the
   socket has been closed
6. Session y's socket thread calls fix_session_deactivate()
7. At this point, the state of the FixSession object is
   up in the air - the main thread could have already started
   freeing the object, which would then give rise to the
   observed segmentation fault

A correct and orderly deactivation of a FixSession object can
be achieved by ensuring the caller of fix_session_deactivate
waits for the Tx, Rx, and socket threads to exit.